feat[close #183]: Implement a update-initramfs command
in some distributions, update-initramfs requires the package manager unlocked to work. In Debian for example it requires dpkg. In my tests even if it was not available, the initramfs was being generated correctly, with some warnings after the process. I think this is not really necessary, maybe convenient. Anyway support is here.